feature: make publishing async
if you want to publish a large number of messages, then publishing them all up-front will probably lead to some of them being dropped by the broker. This change makes publishing async, so that the subscriber can already start emptying the topic at the same time the publisher is filling it up.